Revert "Gg/proxy management"
