+++ src/AppContext.ts
@@ -1,14 +1,6 @@
 import * as React from 'react';
 import { BehaviorSubject, combineLatest, interval, Observable } from 'rxjs';
-import {
-  distinctUntilChanged,
-  filter,
-  first,
-  flatMap,
-  map,
-  shareReplay,
-  switchMap
-} from 'rxjs/operators';
+import { distinctUntilChanged, first, flatMap, map, shareReplay, switchMap } from 'rxjs/operators';
 
 import { isEqual } from 'lodash';
 import { curry } from 'ramda';
@@ -20,9 +12,10 @@ import {
 import {
   Balances,
   CombinedBalances,
+  createAllowances$,
   createBalances$,
   createCombinedBalances$,
-  createDustLimits$, createProxyAllowances$,
+  createDustLimits$,
   createWalletApprove,
   createWalletDisapprove,
   createWethBalances$,
@@ -284,14 +277,7 @@ export function setupAppContext() {
       context$,
       balances$: balancesWithEth$,
       dustLimits$: createDustLimits$(context$),
-      allowances$: createProxyAllowances$(
-        context$,
-        initializedAccount$,
-        proxyAddress$.pipe(
-          filter(address => !!address)
-        ),
-        onEveryBlock$
-      ),
+      allowances$: createAllowances$(context$, initializedAccount$, onEveryBlock$),
     }
   );
