+++ src/blockchain/calls/instant.tsx
@@ -361,7 +361,7 @@ export const setupProxy = {
   options: ({ gasPrice, gasEstimation }: SetupProxyData) =>
     ({ ...gasPrice ? gasPrice : {}, ...gasEstimation ? { gas: gasEstimation } : {} }),
   kind: TxMetaKind.setupProxy,
-  description: () => <React.Fragment>Create Proxy</React.Fragment>
+  description: () => <React.Fragment>Setup proxy</React.Fragment>
 };
 
 export interface SetOwnerData {
