+++ src/instant/Instant.scss.d.ts
@@ -3,7 +3,6 @@ export const header: string;
 export const footer: string;
 export const tradeDetails: string;
 export const details: string;
-export const account: string;
 export const finalization: string;
 export const transaction: string;
 export const assets: string;
@@ -13,7 +12,6 @@ export const hidden: string;
 export const tradingSide: string;
 export const swapIcon: string;
 export const cornerIcon: string;
-export const accountIcon: string;
 export const settingsIcon: string;
 export const backIcon: string;
 export const input: string;
