+++ src/instant/instantForm.test.ts
@@ -11,8 +11,7 @@ import { setupFakeWeb3ForTesting } from '../blockchain/web3';
 import { createFakeOrderbook } from '../exchange/depthChart/fakeOrderBook';
 import { unpack } from '../utils/testHelpers';
 import { zero } from '../utils/zero';
-import { InstantFormChangeKind } from './apply';
-import { createFormController$ } from './instantForm';
+import { createFormController$, InstantFormChangeKind } from './instantForm';
 
 setupFakeWeb3ForTesting();
