+++ src/instant/instantForm.ts
@@ -3,65 +3,118 @@ import { isEqual } from 'lodash';
 import { curry } from 'ramda';
 import { combineLatest, merge, Observable, of, Subject } from 'rxjs';
 import {
-  distinctUntilChanged, first, flatMap,
+  catchError,
+  distinctUntilChanged,
+  first,
+  flatMap,
   map,
   scan,
   shareReplay,
-  switchMap, take,
+  startWith,
+  switchMap,
 } from 'rxjs/operators';
 
 import { Allowances, Balances, DustLimits } from '../balances/balances';
 import { Calls, calls$, Calls$, ReadCalls, ReadCalls$ } from '../blockchain/calls/calls';
-import { NetworkConfig } from '../blockchain/config';
-import { GasPrice$ } from '../blockchain/network';
-import { isDone, isDoneButNotSuccessful, isSuccess, TxState } from '../blockchain/transactions';
+import { eth2weth, weth2eth } from '../blockchain/calls/instant';
+import { NetworkConfig, tokens } from '../blockchain/config';
+import { isDone, TxStatus } from '../blockchain/transactions';
 import { User } from '../blockchain/user';
 import { OfferType } from '../exchange/orderbook/orderbook';
 import { combineAndMerge } from '../utils/combineAndMerge';
 import {
+  BalancesChange,
   doGasEstimation,
+  EtherBalanceChange,
+  EtherPriceUSDChange,
+  FormChangeKind,
   GasEstimationStatus,
+  GasPriceChange,
   HasGasEstimation,
   toBalancesChange,
   toEtherBalanceChange,
   toEtherPriceUSDChange,
   toGasPriceChange,
   toUserChange,
+  UserChange,
 } from '../utils/form';
-import { calculateTradePrice } from '../utils/price';
+import { calculateTradePrice, getQuote } from '../utils/price';
+import { getSlippageLimit } from '../utils/slippage';
 import { switchSpread } from '../utils/switchSpread';
-import {
-  applyChange,
-  InstantFormChangeKind,
-  ManualAllowanceChange,
-  ManualChange,
-  ProgressChange,
-  toAllowancesChange,
-  toContextChange,
-  toDustLimitsChange,
-  toProxyChange
-} from './apply';
-import {
-  evaluateTrade,
-  TradeEvaluationState,
-  TradeEvaluationStatus
-} from './evaluate';
 import { pluginDevModeHelpers } from './instantDevModeHelpers';
 import {
   estimateTradePayWithERC20,
   estimateTradePayWithETH,
-  estimateTradeReadonly, tradePayWithERC20, tradePayWithETH,
+  estimateTradeReadonly,
+  tradePayWithERC20,
+  tradePayWithETH,
 } from './instantTransactions';
-import {
-  ManualAllowanceProgress,
-  ManualAllowanceProgressState,
-  Progress,
-  ProgressKind
-} from './progress/progress';
-import { Message, validate } from './validate';
 
 export interface FormResetChange {
   kind: InstantFormChangeKind.formResetChange;
+}
+
+export enum Position {
+  TOP = 'top',
+  BOTTOM = 'bottom'
+}
+
+export type Placement = Position.TOP | Position.BOTTOM;
+
+export enum MessageKind {
+  noAllowance = 'noAllowance',
+  insufficientAmount = 'insufficientAmount',
+  incredibleAmount = 'incredibleAmount',
+  dustAmount = 'dustAmount',
+  orderbookTotalExceeded = 'orderbookTotalExceeded',
+  notConnected = 'notConnected',
+}
+
+export type Message = {
+  kind: MessageKind.dustAmount
+    | MessageKind.insufficientAmount
+    | MessageKind.incredibleAmount;
+  field: string;
+  priority: number;
+  token: string;
+  amount: BigNumber;
+  placement: Placement;
+} | {
+  kind: MessageKind.orderbookTotalExceeded
+  field: string;
+  side: OfferType
+  amount: BigNumber,
+  token: string;
+  priority: number;
+  placement: Placement;
+  error: any
+} | {
+  kind: MessageKind.notConnected;
+  field: string;
+  priority: number;
+  placement: Placement;
+// } | {
+//   kind: MessageKind.custom
+//   field: string;
+//   priority: number;
+//   placement: Placement;
+//   error: any
+};
+
+export enum TradeEvaluationStatus {
+  unset = 'unset',
+  calculating = 'calculating',
+  calculated = 'calculated',
+  error = 'error',
+}
+
+export enum ProgressKind {
+  onlyProxy = 'onlyProxy',
+  proxyPayWithETH = 'proxyPayWithETH',
+  noProxyPayWithETH = 'noProxyPayWithETH',
+  noProxyNoAllowancePayWithERC20 = 'noProxyNoAllowancePayWithERC20',
+  proxyNoAllowancePayWithERC20 = 'proxyNoAllowancePayWithERC20',
+  proxyAllowancePayWithERC20 = 'proxyAllowancePayWithERC20',
 }
 
 export enum ViewKind {
@@ -76,9 +129,46 @@ export enum ViewKind {
   summary = 'summary'
 }
 
-export interface InstantFormState extends HasGasEstimation,
-  TradeEvaluationState,
-  ManualAllowanceProgressState {
+export type Progress = {
+  gasUsed?: BigNumber,
+  bought?: BigNumber,
+  sold?: BigNumber
+  done: boolean;
+} & ({
+  kind: ProgressKind.proxyPayWithETH
+    | ProgressKind.noProxyPayWithETH
+    | ProgressKind.proxyAllowancePayWithERC20
+  tradeTxStatus: TxStatus;
+  tradeTxHash?: string;
+} | {
+  kind: ProgressKind.noProxyNoAllowancePayWithERC20
+  proxyTxStatus: TxStatus;
+  proxyTxHash?: string
+  allowanceTxStatus?: TxStatus;
+  allowanceTxHash?: string
+  tradeTxStatus?: TxStatus;
+  tradeTxHash?: string
+} | {
+  kind: ProgressKind.proxyNoAllowancePayWithERC20;
+  allowanceTxStatus: TxStatus;
+  allowanceTxHash?: string
+  tradeTxStatus?: TxStatus;
+  tradeTxHash?: string
+} | {
+  kind: ProgressKind.onlyProxy
+  proxyTxStatus: TxStatus;
+  tradeTxStatus?: TxStatus;
+});
+
+interface TradeEvaluationState {
+  buyAmount?: BigNumber;
+  sellAmount?: BigNumber;
+  tradeEvaluationStatus: TradeEvaluationStatus;
+  tradeEvaluationError?: any;
+  bestPrice?: BigNumber;
+}
+
+export interface InstantFormState extends HasGasEstimation, TradeEvaluationState {
   view: ViewKind;
   readyToProceed?: boolean;
   progress?: Progress;
@@ -100,6 +190,343 @@ export interface InstantFormState extends HasGasEstimation,
   proxyAddress?: string;
   user?: User;
   context?: NetworkConfig;
+}
+
+export enum InstantFormChangeKind {
+  viewChange = 'viewChange',
+  tokenChange = 'tokenChange',
+  buyAmountFieldChange = 'buyAmountFieldChange',
+  sellAmountFieldChange = 'sellAmountFieldChange',
+  pairChange = 'pairChange',
+  formResetChange = 'reset',
+  progressChange = 'progressChange',
+  proxyChange = 'proxyChange',
+  dustLimitsChange = 'dustLimitsChange',
+  allowancesChange = 'allowancesChange',
+  slippageLimitChange = 'slippageLimitChange',
+  contextChange = 'contextChange',
+}
+
+export interface ProgressChange {
+  kind: InstantFormChangeKind.progressChange;
+  progress?: Progress;
+}
+
+export interface ViewChange {
+  kind: InstantFormChangeKind.viewChange;
+  view: ViewKind;
+  meta?: any;
+}
+
+export interface BuyAmountChange {
+  kind: InstantFormChangeKind.buyAmountFieldChange;
+  value?: BigNumber;
+}
+
+export interface SellAmountChange {
+  kind: InstantFormChangeKind.sellAmountFieldChange;
+  value?: BigNumber;
+}
+
+export interface PairChange {
+  kind: InstantFormChangeKind.pairChange;
+  buyToken: string;
+  sellToken: string;
+}
+
+export interface TokenChange {
+  kind: InstantFormChangeKind.tokenChange;
+  token: string;
+  side: OfferType;
+}
+
+export interface DustLimitsChange {
+  kind: InstantFormChangeKind.dustLimitsChange;
+  dustLimits: DustLimits;
+}
+
+export interface AllowancesChange {
+  kind: InstantFormChangeKind.allowancesChange;
+  allowances: Allowances;
+}
+
+export interface ProxyChange {
+  kind: InstantFormChangeKind.proxyChange;
+  value?: string;
+}
+
+export interface ContextChange {
+  kind: InstantFormChangeKind.contextChange;
+  context: NetworkConfig;
+}
+
+export interface SlippageLimitChange {
+  kind: InstantFormChangeKind.slippageLimitChange;
+  value: BigNumber;
+}
+
+export type ManualChange =
+  BuyAmountChange |
+  SellAmountChange |
+  PairChange |
+  TokenChange |
+  FormResetChange |
+  ViewChange |
+  SlippageLimitChange;
+
+export type EnvironmentChange =
+  GasPriceChange |
+  EtherPriceUSDChange |
+  BalancesChange |
+  DustLimitsChange |
+  AllowancesChange |
+  ProxyChange |
+  EtherBalanceChange |
+  UserChange |
+  ContextChange;
+
+export type InstantFormChange =
+  ManualChange |
+  EnvironmentChange |
+  ProgressChange;
+
+function applyChange(state: InstantFormState, change: InstantFormChange): InstantFormState {
+  switch (change.kind) {
+    case InstantFormChangeKind.pairChange:
+      return {
+        ...state,
+        buyToken: change.buyToken,
+        sellToken: change.sellToken,
+        kind: undefined,
+        buyAmount: undefined,
+        sellAmount: undefined,
+        slippageLimit: state.context
+          ? getSlippageLimit(
+            state.context,
+            getQuote(weth2eth(state.sellToken), weth2eth(state.buyToken))
+          )
+          : state.slippageLimit,
+        tradeEvaluationStatus: TradeEvaluationStatus.unset,
+      };
+    case InstantFormChangeKind.tokenChange:
+      const { side, token } = change;
+      const currentBuyToken = state.buyToken;
+      const currentSellToken = state.sellToken;
+
+      let buyToken = currentBuyToken;
+      let sellToken = currentSellToken;
+      let shouldClearInputs = false;
+      if (side === OfferType.sell) {
+        sellToken = token;
+        shouldClearInputs = token !== currentSellToken;
+      }
+
+      if (side === OfferType.buy) {
+        buyToken = token;
+        shouldClearInputs = token !== currentBuyToken;
+      }
+
+      if (side === OfferType.buy && eth2weth(token) === eth2weth(currentSellToken)) {
+        buyToken = token;
+        sellToken = currentBuyToken;
+        shouldClearInputs = true;
+      }
+
+      if (side === OfferType.sell && eth2weth(token) === eth2weth(currentBuyToken)) {
+        buyToken = currentSellToken;
+        sellToken = token;
+        shouldClearInputs = true;
+      }
+
+      return {
+        ...state,
+        buyToken,
+        sellToken,
+        buyAmount: shouldClearInputs ? undefined : state.buyAmount,
+        sellAmount: shouldClearInputs ? undefined : state.sellAmount,
+        slippageLimit: state.context
+          ? getSlippageLimit(
+            state.context,
+            getQuote(weth2eth(state.sellToken), weth2eth(state.buyToken))
+          )
+          : state.slippageLimit,
+        tradeEvaluationStatus: TradeEvaluationStatus.unset,
+      };
+    case InstantFormChangeKind.sellAmountFieldChange:
+      return {
+        ...state,
+        kind: OfferType.sell,
+        sellAmount: change.value,
+      };
+    case InstantFormChangeKind.buyAmountFieldChange:
+      return {
+        ...state,
+        kind: OfferType.buy,
+        buyAmount: change.value,
+      };
+    case FormChangeKind.gasPriceChange:
+      return {
+        ...state,
+        gasPrice: change.value,
+      };
+    case FormChangeKind.etherPriceUSDChange:
+      return {
+        ...state,
+        etherPriceUsd: change.value,
+      };
+    case FormChangeKind.balancesChange:
+      return {
+        ...state,
+        balances: change.balances,
+      };
+    case InstantFormChangeKind.dustLimitsChange:
+      return {
+        ...state,
+        dustLimits: change.dustLimits
+      };
+    case InstantFormChangeKind.allowancesChange:
+      return {
+        ...state,
+        allowances: change.allowances
+      };
+    case InstantFormChangeKind.progressChange:
+
+      if (change.progress) {
+        return {
+          ...state,
+          progress: change.progress,
+          view: change.progress.tradeTxStatus === TxStatus.Success
+            ? ViewKind.summary
+            : ViewKind.finalization
+        };
+      }
+
+      return state;
+
+    case InstantFormChangeKind.formResetChange:
+      return {
+        ...state,
+        progress: undefined,
+        buyAmount: undefined,
+        sellAmount: undefined,
+        gasEstimationStatus: GasEstimationStatus.unset
+      };
+    case FormChangeKind.etherBalanceChange:
+      return {
+        ...state,
+        etherBalance: change.etherBalance
+      };
+    case InstantFormChangeKind.viewChange:
+      return {
+        ...state,
+        view: change.view
+      };
+    case InstantFormChangeKind.proxyChange:
+      return {
+        ...state,
+        proxyAddress: change.value
+      };
+    case InstantFormChangeKind.contextChange:
+      return {
+        ...state,
+        context: change.context,
+        slippageLimit: getSlippageLimit(
+          change.context,
+          getQuote(weth2eth(state.sellToken), weth2eth(state.buyToken))
+        )
+      };
+    case FormChangeKind.userChange:
+      return {
+        ...state,
+        user: change.user
+      };
+    case InstantFormChangeKind.slippageLimitChange:
+      return {
+        ...state,
+        slippageLimit: change.value
+      };
+  }
+
+  return state;
+}
+
+function evaluateBuy(calls: ReadCalls, state: InstantFormState) {
+
+  const { buyToken, sellToken, buyAmount } = state;
+
+  if (!buyToken || !sellToken || !buyAmount) {
+    return of(state);
+  }
+
+  const errorItem = (error?: Error) => ({
+    sellAmount: undefined,
+    message: {
+      error,
+      kind: MessageKind.orderbookTotalExceeded,
+      amount: buyAmount,
+      side: 'buy',
+      token: buyToken,
+      placement: Position.TOP,
+      priority: 3
+    }
+  });
+
+  return calls.otcGetPayAmount({
+    sellToken,
+    buyToken,
+    amount: buyAmount,
+  }).pipe(
+    switchMap(sellAmount => of(sellAmount.isZero() ? errorItem() : { sellAmount })),
+    catchError(error => of(errorItem(error))),
+  );
+}
+
+function evaluateSell(calls: ReadCalls, state: InstantFormState) {
+
+  const { buyToken, sellToken, sellAmount } = state;
+
+  if (!buyToken || !sellToken || !sellAmount) {
+    return of(state);
+  }
+
+  const errorItem = (error?: Error) => ({
+    buyAmount: undefined,
+    message: {
+      error,
+      kind: MessageKind.orderbookTotalExceeded,
+      amount: sellAmount,
+      side: 'sell',
+      token: sellToken,
+      placement: Position.TOP,
+      priority: 3
+    }
+  });
+
+  return calls.otcGetBuyAmount({
+    sellToken,
+    buyToken,
+    amount: sellAmount,
+  }).pipe(
+    switchMap(buyAmount => of(buyAmount.isZero() ? errorItem() : { buyAmount })),
+    catchError(error => of(errorItem(error))),
+  );
+}
+
+function getBestPrice(
+  calls: ReadCalls,
+  sellToken: string,
+  buyToken: string
+): Observable<BigNumber> {
+  return calls.otcGetBestOffer({ sellToken, buyToken }).pipe(
+    flatMap(offerId =>
+      calls.otcOffers(offerId).pipe(
+        map(([a, _, b]: BigNumber[]) => {
+          return (sellToken === 'DAI' || (sellToken === 'WETH' && buyToken !== 'DAI')) ?
+            a.div(b) : b.div(a);
+        })
+      )
+    )
+  );
 }
 
 function estimateGas(calls$_: Calls$, readCalls$: ReadCalls$, state: InstantFormState) {
@@ -133,6 +560,63 @@ function gasEstimation(
     );
 }
 
+function evaluateTrade(
+  theCalls$: ReadCalls$, previousState: InstantFormState, state: InstantFormState
+): Observable<TradeEvaluationState> | undefined {
+
+  if (
+    state.kind === OfferType.buy &&
+    state.kind === previousState.kind &&
+    state.buyAmount && previousState.buyAmount &&
+    state.buyAmount.eq(previousState.buyAmount)
+    ||
+    state.kind === OfferType.sell &&
+    state.kind === previousState.kind &&
+    state.sellAmount && previousState.sellAmount &&
+    state.sellAmount.eq(previousState.sellAmount)
+  ) {
+    return undefined;
+  }
+
+  if (
+    !state.kind || state.kind === OfferType.buy && !state.buyAmount
+    || state.buyAmount && state.buyAmount.eq(new BigNumber(0))
+    || state.kind === OfferType.sell && !state.sellAmount
+    || state.sellAmount && state.sellAmount.eq(new BigNumber(0))
+  ) {
+    return of({
+      tradeEvaluationStatus: TradeEvaluationStatus.unset,
+      ...state.kind === OfferType.buy ? { sellAmount: undefined } : { buyAmount: undefined }
+    });
+  }
+
+  return theCalls$.pipe(
+    first(),
+    switchMap(calls =>
+      combineLatest(
+        state.kind === OfferType.buy ? evaluateBuy(calls, state) : evaluateSell(calls, state),
+        // tslint:disable-next-line:max-line-length
+        // This is some suspicious case. This way it works like we had on OD but needs in-depth investigation.
+        getBestPrice(calls, state.buyToken, state.sellToken)
+      )
+    ),
+    map(([evaluation, bestPrice]) => ({
+      ...evaluation,
+      bestPrice,
+      tradeEvaluationStatus: TradeEvaluationStatus.calculated,
+    })),
+    startWith({ ...state, tradeEvaluationStatus: TradeEvaluationStatus.calculating }),
+    catchError(err => {
+      return of({
+        ...(state.kind === OfferType.buy ? { sellAmount: undefined } : { buyAmount: undefined }),
+        bestPrice: undefined,
+        tradeEvaluationStatus: TradeEvaluationStatus.error,
+        tradeEvaluationError: err,
+      });
+    }),
+  );
+}
+
 function mergeTradeEvaluation(
   state: InstantFormState, trade: TradeEvaluationState | undefined
 ): InstantFormState {
@@ -142,6 +626,103 @@ function mergeTradeEvaluation(
   return {
     ...state,
     ...trade
+  };
+}
+
+function validate(state: InstantFormState): InstantFormState {
+  if (state.tradeEvaluationStatus !== TradeEvaluationStatus.calculated) {
+    return state;
+  }
+
+  let message: Message | undefined = state.message;
+
+  const [spendField, receiveField] = ['sellToken', 'buyToken'];
+  const [spendToken, receiveToken] = [state.sellToken, state.buyToken];
+  const [spendAmount, receiveAmount] = [state.sellAmount, state.buyAmount];
+  const dustLimits = state.dustLimits;
+
+  if (!state.user || !state.user.account) {
+    message = prioritize(message, {
+      kind: MessageKind.notConnected,
+      field: spendField,
+      priority: 1000,
+      placement: Position.BOTTOM,
+    });
+  }
+
+  if (spendAmount && (
+    spendToken === 'ETH' && state.etherBalance && state.etherBalance.lt(spendAmount) ||
+    state.balances && state.balances[spendToken] && state.balances[spendToken].lt(spendAmount)
+  )) {
+    message = prioritize(message, {
+      kind: MessageKind.insufficientAmount,
+      field: spendField,
+      amount: spendAmount,
+      priority: 1,
+      token: spendToken,
+      placement: Position.BOTTOM
+    });
+  }
+
+  if (
+    spendAmount
+    && dustLimits
+    && dustLimits[eth2weth(spendToken)]
+    && dustLimits[eth2weth(spendToken)].gt(spendAmount)
+  ) {
+    message = prioritize(message, {
+      kind: MessageKind.dustAmount,
+      amount: dustLimits[eth2weth(spendToken)],
+      field: spendField,
+      priority: 2,
+      token: spendToken,
+      placement: Position.BOTTOM
+    });
+  }
+
+  if (
+    receiveAmount
+    && dustLimits
+    && dustLimits[eth2weth(receiveToken)]
+    && dustLimits[eth2weth(receiveToken)].gt(receiveAmount)
+  ) {
+    message = prioritize(message, {
+      kind: MessageKind.dustAmount,
+      amount: dustLimits[eth2weth(receiveToken)],
+      field: receiveField,
+      priority: 2,
+      token: receiveToken,
+      placement: Position.BOTTOM
+    });
+  }
+
+  if (
+    spendAmount && new BigNumber(tokens[eth2weth(spendToken)].maxSell).lt(spendAmount)
+  ) {
+    message = prioritize(message, {
+      kind: MessageKind.incredibleAmount,
+      field: spendField,
+      priority: 2,
+      token: spendToken,
+      amount: new BigNumber(tokens[eth2weth(spendToken)].maxSell),
+      placement: Position.BOTTOM
+    });
+  }
+
+  if (receiveAmount && new BigNumber(tokens[eth2weth(receiveToken)].maxSell).lt(receiveAmount)) {
+    message = prioritize(message, {
+      kind: MessageKind.incredibleAmount,
+      field: receiveField,
+      priority: 2,
+      token: receiveToken,
+      amount: new BigNumber(tokens[eth2weth(receiveToken)].maxSell),
+      placement: Position.BOTTOM
+    });
+  }
+
+  return {
+    ...state,
+    message
   };
 }
 
@@ -168,150 +749,7 @@ function calculatePriceAndImpact(state: InstantFormState): InstantFormState {
   };
 }
 
-function isReadyToProceed(state: InstantFormState): InstantFormState {
-  return {
-    ...state,
-    readyToProceed: !state.message && state.gasEstimationStatus === GasEstimationStatus.calculated
-  };
-}
-
-function freezeIfInProgress(previous: InstantFormState, state: InstantFormState): InstantFormState {
-  if (state.view === ViewKind.finalization) {
-    return {
-      ...previous,
-      progress: state.progress,
-      view: state.view
-    };
-  }
-  return state;
-}
-
-enum AllowanceDirection {
-  locking = 'locking',
-  unlocking = 'unlocking',
-}
-
-interface UnidirectionalManualAllowanceStatus {
-  token: string;
-  direction: AllowanceDirection;
-  progress: TxState;
-}
-
-export function manualAllowanceSetup(
-  theCalls$: Calls$,
-  gasPrice$: GasPrice$,
-  proxyAddress$: Observable<string>,
-  allowances$: Observable<Allowances>
-): [(token: string) => void, Observable<ManualAllowanceChange>] {
-  const manualAllowanceProgressChanges$ = new Subject<ManualAllowanceChange>();
-  const allowanceStatus$ = new Subject<UnidirectionalManualAllowanceStatus>();
-
-  function toggleAllowance(token: string) {
-    theCalls$.pipe(
-      first(),
-      switchMap((calls) =>
-        combineLatest(proxyAddress$, allowances$).pipe(
-          take(1),
-          switchMap(([proxyAddress, allowances]) =>
-            combineLatest(calls.approveProxyEstimateGas({ token, proxyAddress }), gasPrice$).pipe(
-              take(1),
-              switchMap(([estimation, gasPrice]) => {
-                const gasCost = {
-                  gasPrice,
-                  gasEstimation: estimation,
-                };
-
-                return allowances[token]
-                  ? calls.disapproveProxy({ proxyAddress, token, ...gasCost }).pipe(
-                    switchMap(progress => of({
-                      token,
-                      progress,
-                      direction: AllowanceDirection.locking,
-                    }))
-                  )
-                  : calls.approveProxy({ proxyAddress, token, ...gasCost }).pipe(
-                    switchMap(progress => of({
-                      token,
-                      progress,
-                      direction: AllowanceDirection.unlocking
-                    }))
-                  );
-              })
-            )
-          )
-        )
-      ),
-    ).subscribe((txProgress) => {
-      allowanceStatus$.next(txProgress);
-    });
-  }
-
-  combineLatest(allowances$, allowanceStatus$).subscribe(
-    ([allowances, allowanceStatus]) => {
-      const { token, direction, progress } = allowanceStatus;
-      manualAllowanceProgressChanges$.next({
-        token,
-        kind: InstantFormChangeKind.manualAllowanceChange,
-        progress: {
-          token,
-          direction,
-          kind: ProgressKind.onlyAllowance,
-          allowanceTxStatus: progress.status,
-          txHash: (progress as { txHash: string; }).txHash,
-          done: isSuccess(progress) && (
-            // If we are unlocking the given token, we wait until it's
-            // allowed which will be visible on the next block check.
-            (direction === 'unlocking' && allowances[token])
-            // If we are locking the given token, we wait until it's
-            // not allowed which will be visible on the next block check.
-            || (direction === 'locking' && !allowances[token])
-          ) || isDoneButNotSuccessful(progress)
-        } as ManualAllowanceProgress
-      });
-    }
-  );
-
-  return [toggleAllowance, manualAllowanceProgressChanges$];
-}
-
-export function manualProxyCreation(
-  theCalls$: Calls$,
-  gasPrice$: GasPrice$,
-): [() => void, Observable<ProgressChange>] {
-
-  const proxyCreationChange$ = new Subject<ProgressChange>();
-
-  function createProxy() {
-    theCalls$.pipe(
-      first(),
-      switchMap(calls =>
-        combineLatest(calls.setupProxyEstimateGas({}), gasPrice$)
-          .pipe(
-            switchMap(([estimatedGas, gasPrice]) =>
-              calls.setupProxy({
-                gasPrice,
-                gasEstimation: estimatedGas
-              })
-            )
-          )
-      ),
-    ).subscribe(progress => {
-      proxyCreationChange$.next({
-        kind: InstantFormChangeKind.progressChange,
-        progress: {
-          kind: ProgressKind.onlyProxy,
-          proxyTxStatus: progress.status,
-          txHash: (progress as { txHash: string; }).txHash,
-          done: isDone(progress)
-        }
-      });
-    });
-  }
-
-  return [createProxy, proxyCreationChange$];
-}
-
-export function prepareSubmit(
+function prepareSubmit(
   theCalls$: Calls$,
 ): [(state: InstantFormState) => void, Observable<ProgressChange | FormResetChange>] {
 
@@ -331,6 +769,86 @@ export function prepareSubmit(
   }
 
   return [submit, stageChange$];
+}
+
+function manualProxyCreation(
+  theCalls$: Calls$,
+): [() => void, Observable<ProgressChange>] {
+
+  const proxyCreationChange$ = new Subject<ProgressChange>();
+
+  function createProxy() {
+    theCalls$.pipe(
+      flatMap((calls) =>
+        calls.setupProxy({})
+      )
+    ).subscribe(progress => {
+      proxyCreationChange$.next({
+        kind: InstantFormChangeKind.progressChange,
+        progress: {
+          kind: ProgressKind.onlyProxy,
+          proxyTxStatus: progress.status,
+          done: isDone(progress)
+        }
+      });
+    });
+  }
+
+  return [createProxy, proxyCreationChange$];
+}
+
+function toDustLimitsChange(dustLimits$: Observable<DustLimits>): Observable<DustLimitsChange> {
+  return dustLimits$.pipe(
+    map(dustLimits => ({
+      dustLimits,
+      kind: InstantFormChangeKind.dustLimitsChange,
+    } as DustLimitsChange))
+  );
+}
+
+function toAllowancesChange(allowances$: Observable<Allowances>): Observable<AllowancesChange> {
+  return allowances$.pipe(
+    map(allowances => ({
+      allowances,
+      kind: InstantFormChangeKind.allowancesChange,
+    } as AllowancesChange))
+  );
+}
+
+function toProxyChange(proxyAddress$: Observable<string>): Observable<ProxyChange> {
+  return proxyAddress$.pipe(
+    map(proxy => ({
+      value: proxy,
+      kind: InstantFormChangeKind.proxyChange,
+    } as ProxyChange))
+  );
+}
+
+function toContextChange(context$: Observable<NetworkConfig>): Observable<ContextChange> {
+  return context$.pipe(
+    map(context => ({
+      context,
+      kind: InstantFormChangeKind.contextChange,
+    } as ContextChange))
+  );
+}
+
+function isReadyToProceed(state: InstantFormState): InstantFormState {
+  return {
+    ...state,
+    readyToProceed: !state.message && state.gasEstimationStatus === GasEstimationStatus.calculated
+  };
+}
+
+function freezeIfInProgress(previous: InstantFormState, state: InstantFormState): InstantFormState {
+  if (state.progress || previous.view !== state.view) {
+    return {
+      ...previous,
+      progress: state.progress,
+      view: state.view
+    };
+  }
+  return state;
 }
 
 export function createFormController$(
@@ -368,22 +886,11 @@ export function createFormController$(
   );
 
   const [submit, submitChange$] = prepareSubmit(params.calls$);
-
-  const [createProxy, createProxyManuallyChange$] =
-    manualProxyCreation(params.calls$, params.gasPrice$);
-
-  const [toggleAllowance, setAllowanceManuallyChange$] =
-    manualAllowanceSetup(
-      params.calls$,
-      params.gasPrice$,
-      params.proxyAddress$,
-      params.allowances$
-    );
+  const [createProxy, proxyCreationChange$] = manualProxyCreation(params.calls$);
 
   const initialState: InstantFormState = {
     submit,
     createProxy,
-    toggleAllowance,
     change: manualChange$.next.bind(manualChange$),
     buyToken: 'DAI',
     sellToken: 'ETH',
@@ -401,8 +908,7 @@ export function createFormController$(
   return merge(
     manualChange$,
     submitChange$,
-    createProxyManuallyChange$,
-    setAllowanceManuallyChange$,
+    proxyCreationChange$,
     environmentChange$,
   ).pipe(
     scan(applyChange, initialState),
@@ -419,3 +925,17 @@ export function createFormController$(
     shareReplay(1),
   );
 }
+
+const prioritize = (current: Message = { priority: 0 } as Message, candidate: Message) => {
+  // Prioritize by priority first
+  if (current.priority < candidate.priority) {
+    return candidate;
+  }
+
+  // and if we have errors with same priority, the one for paying input is more important
+  if (current.priority === candidate.priority) {
+    return current.field === 'sellToken' ? current : candidate;
+  }
+
+  return current;
+};
