+++ src/instant/views/AccountView.scss
@@ -21,13 +21,6 @@
 }
 
 .proxy {
-  &Missing, &Available {
-    .text {
-      display: inline-block;
-      margin-right: 6px;
-    }
-  }
-
   &Missing {
     .text {
       color: white;
@@ -74,14 +67,14 @@
   }
 
   .warningIcon {
-    height: 1.5rem;
-    width: 1.5rem;
+    height: 1.375rem;
+    width: 1.375rem;
     flex-shrink: 0;
     margin-right: .8125rem;
     margin-left: 0;
 
     :global(.warning-icon) {
-      fill: white;
+      fill: $grey-lighter;
     }
   }
 }
