+++ src/instant/views/AccountView.scss.d.ts
@@ -1,9 +1,9 @@
 export const row: string;
 export const proxyMissing: string;
 export const text: string;
-export const proxyAvailable: string;
 export const accountIcon: string;
 export const warningIcon: string;
+export const proxyAvailable: string;
 export const warning: string;
 export const allowances: string;
 export const button: string;
