+++ src/instant/views/AccountView.tsx
@@ -1,7 +1,6 @@
 import classnames from 'classnames';
 import * as React from 'react';
 import { Allowances } from '../../balances/balances';
-import { TxStatus } from '../../blockchain/transactions';
 import accountSvg from '../../icons/account.svg';
 import backSvg from '../../icons/back.svg';
 import warningSvg from '../../icons/warning.svg';
@@ -9,24 +8,10 @@ import { Button } from '../../utils/forms/Buttons';
 import { ButtonIcon, ProgressIcon } from '../../utils/icons/Icons';
 import { SvgImage } from '../../utils/icons/utils';
 import { TopLeftCorner } from '../../utils/panel/TopRightCorner';
-import { WarningTooltip, WarningTooltipType } from '../../utils/tooltip/Tooltip';
-import { InstantFormChangeKind, ManualChange } from '../apply';
 import * as instantStyles from '../Instant.scss';
-import { ViewKind } from '../instantForm';
+import { InstantFormChangeKind, ManualChange, Progress, ViewKind } from '../instantForm';
 import { InstantFormWrapper } from '../InstantFormWrapper';
-import { Progress } from '../progress/progress';
 import * as styles from './AccountView.scss';
-
-// tslint:disable
-const proxyToolTip = {
-  id: 'proxy-tooltip',
-  text: 'Proxy is a supporting contract owned by you that groups different actions as one Ethereum transaction.',
-  iconColor: 'white'
-} as WarningTooltipType;
-
-const activeProxyTooltip = { ...proxyToolTip, iconColor: 'soft-cyan' as 'soft-cyan' };
-
-// tslint:enable
 
 export interface ViewProps {
   proxyAddress: string;
@@ -43,7 +28,7 @@ const box = {
 
 export class AccountView extends React.Component<ViewProps> {
   public render() {
-    const { proxyAddress, progress } = this.props;
+    const { proxyAddress } = this.props;
 
     return (
       <InstantFormWrapper heading={'Account Overview'}>
@@ -54,14 +39,9 @@ export class AccountView extends React.Component<ViewProps> {
             image={backSvg}
           />
         </TopLeftCorner>
-        <div className={classnames(instantStyles.details, instantStyles.account)}>
+        <div className={instantStyles.details}>
           {
-            proxyAddress || (
-              proxyAddress
-              && progress
-              && progress.done
-              && (progress as { proxyTxStatus: TxStatus }).proxyTxStatus === 'Success'
-            )
+            proxyAddress
               ? this.onHavingProxy()
               : this.onMissingProxy()
           }
@@ -76,14 +56,14 @@ export class AccountView extends React.Component<ViewProps> {
         <div style={box}>
           <SvgImage className={styles.accountIcon} image={accountSvg}/>
           <span className={styles.text}>Proxy available</span>
-          <WarningTooltip {...activeProxyTooltip}/>
+          <SvgImage className={styles.warningIcon} image={warningSvg}/>
         </div>
       </div>
       <div className={classnames(styles.row, styles.allowances)}>
         <span className={styles.text}>{this.allowedTokens()} Tokens enabled for Trading</span>
         <Button
           size="sm"
-          color="grey"
+          color="greyWhite"
           className={styles.button}
           onClick={this.switchToAllowances}
         >
@@ -95,41 +75,41 @@ export class AccountView extends React.Component<ViewProps> {
 
   private onMissingProxy = () => {
     const progress = this.props.progress;
-    const isInProgress = progress && !progress.done && !this.props.proxyAddress;
+    const isInProgress = progress && !progress.done;
     return (
-      <>
-        <div className={classnames(styles.row, styles.proxyMissing)}>
-          <div style={box}>
-            <SvgImage className={styles.accountIcon} image={accountSvg}/>
-            <span className={styles.text}>Proxy not created</span>
-            <WarningTooltip {...proxyToolTip}/>
-          </div>
-          <div className={styles.placeholder}>
-            {
-              isInProgress
-                ? <ProgressIcon/>
-                : (
-                  <Button
-                    size="sm"
-                    color="grey"
-                    className={styles.button}
-                    onClick={this.props.createProxy}
-                  >
-                    Create
-                  </Button>
-                )
-            }
-          </div>
+    <>
+      <div className={classnames(styles.row, styles.proxyMissing)}>
+        <div style={box}>
+          <SvgImage className={styles.accountIcon} image={accountSvg}/>
+          <span className={styles.text}>Proxy not created</span>
+          <SvgImage className={styles.warningIcon} image={warningSvg}/>
         </div>
-        <div className={classnames(styles.row, styles.warning)}>
-          <>
-            <SvgImage className={styles.warningIcon} image={warningSvg}/>
-            <span className={styles.text}>
+        <div className={styles.placeholder}>
+          {
+            isInProgress
+              ? <ProgressIcon/>
+              : (
+                <Button
+                  size="sm"
+                  color="greyWhite"
+                  className={styles.button}
+                  onClick={this.props.createProxy}
+                >
+                  Create
+                </Button>
+              )
+          }
+        </div>
+      </div>
+      <div className={classnames(styles.row, styles.warning)}>
+        <>
+          <SvgImage className={styles.warningIcon} image={warningSvg}/>
+          <span className={styles.text}>
             You do not need to create a proxy manually. It will be automatically created for you.
           </span>
-          </>
-        </div>
-      </>
+        </>
+      </div>
+    </>
     );
   }
 
@@ -145,7 +125,7 @@ export class AccountView extends React.Component<ViewProps> {
 
   private allowedTokens = () => {
     return Object
-      .values(this.props.allowances || [])
+      .values(this.props.allowances)
       .reduce((allowedTokensCount: number, isAllowed: boolean) => {
         return isAllowed ? allowedTokensCount + 1 : allowedTokensCount;
       },      0);
