+++ src/instant/views/AllowancesView.scss
@@ -1,30 +1,23 @@
 @import "../../utils/Variables";
-@import "../../utils/Mixins";
 
-.assets {
+.list {
   display: flex;
   align-items: center;
   justify-content: flex-start;
   flex-wrap: wrap;
   list-style-type: none;
   padding: 0;
-  margin-top: 1rem;
-
-  @include media-breakpoint-down(md){
-    justify-content: center;
-  }
 }
 
-.asset {
+.listItem {
   position: relative;
   display: inline-flex;
   align-items: center;
-  border-radius: 4px;
+  border: 1px solid #3F3F47;
   padding: 1rem 1.5rem;
   margin: .25rem;
-  font-size: 16px;
-  font-weight: 500;
-  letter-spacing: .2px;
+  font-size: 12px;
+  font-weight: 400;
   width: 100%;
   max-width: 212px;
 }
@@ -35,7 +28,7 @@
   margin-right: 12px;
 }
 
-.indicator {
+.doneIcon {
   position: absolute;
   top: 10px;
   right: 10px;
@@ -45,12 +38,6 @@
 
 .isAllowed {
   :global(.done-icon) {
-    fill: white;
-  }
-}
-
-.disabled {
-  :global(.done-icon) {
-    fill: $yet-another-dark-grayish-blue;
+    fill: $soft-cyan;
   }
 }
