+++ src/instant/views/AllowancesView.scss.d.ts
@@ -1,6 +1,5 @@
-export const assets: string;
-export const asset: string;
+export const list: string;
+export const listItem: string;
 export const tokenIcon: string;
-export const indicator: string;
+export const doneIcon: string;
 export const isAllowed: string;
-export const disabled: string;
