+++ src/instant/views/AllowancesView.tsx
@@ -1,100 +1,53 @@
+import classnames from 'classnames';
 import * as React from 'react';
 import { Allowances } from '../../balances/balances';
 import { tokens } from '../../blockchain/config';
 import doneSvg from '../../icons/done.svg';
-import { Button, CloseButton } from '../../utils/forms/Buttons';
+import { CloseButton } from '../../utils/forms/Buttons';
 import { SvgImage } from '../../utils/icons/utils';
-import { LoadingIndicator } from '../../utils/loadingIndicator/LoadingIndicator';
 import { TopRightCorner } from '../../utils/panel/TopRightCorner';
-import { InstantFormChangeKind, ManualChange } from '../apply';
-import * as instantStyles from '../Instant.scss';
-import {
-  ViewKind
-} from '../instantForm';
+import { InstantFormChangeKind, ManualChange, ViewKind } from '../instantForm';
 import { InstantFormWrapper } from '../InstantFormWrapper';
-import { ManualAllowanceProgressState } from '../progress/progress';
 import * as styles from './AllowancesView.scss';
 
-interface AssetProps {
-  isAllowed: boolean;
-  asset: any;
-  inProgress?: boolean;
-  onClick: () => void;
-}
-
-class AssetAllowance extends React.Component<AssetProps> {
-  public render() {
-    // @ts-ignore
-    const { isAllowed, asset, inProgress, onClick } = this.props;
-
-    return (
-      <Button color="grey"
-              disabled={inProgress}
-              className={styles.asset}
-              onClick={onClick}
-      >
-        <span className={styles.tokenIcon}>{asset.iconColor}</span>
-        <span>{asset.symbol}</span>
-        <span className={styles.indicator}>
-          {
-            inProgress
-              ? <LoadingIndicator inline={true}/>
-              : <SvgImage className={
-                isAllowed ? styles.isAllowed : styles.disabled
-              } image={doneSvg}
-              />
-          }
-        </span>
-      </Button>
-    );
-  }
-}
-
-interface ViewProps extends ManualAllowanceProgressState {
+interface ViewProps {
   allowances: Allowances;
   change: (change: ManualChange) => void;
 }
 
 export class AllowancesView extends React.Component<ViewProps> {
   public render() {
-    const { allowances, toggleAllowance, manualAllowancesProgress } = this.props;
+    const allowances = this.props.allowances;
 
     return (
-      <InstantFormWrapper heading={'Unlock Token for Trading'}>
+      <InstantFormWrapper heading={'Enable Token for Trading'}>
         <TopRightCorner>
-          <CloseButton theme="danger"
-                       className={instantStyles.closeButton}
-                       onClick={this.close}
-          />
+          <CloseButton onClick={this.onClose}/>
         </TopRightCorner>
-        <div className={styles.assets}>
+        <ul className={styles.list}>
           {
             Object.values(tokens)
-              .filter(token => token.symbol !== 'ETH')
-              .map(
-                (token: any, index: number) => {
-                  const symbol = token.symbol;
-                  const progress = manualAllowancesProgress && manualAllowancesProgress[symbol];
-
-                  return <AssetAllowance isAllowed={allowances[symbol]}
-                                         inProgress={progress && !progress.done}
-                                         key={index}
-                                         asset={token}
-                                         onClick={() => {
-                                           toggleAllowance(symbol);
-                                         }}
-                  />;
-                }
-              )
+              .filter(asset => asset.symbol !== 'ETH')
+              .map((asset, index) => {
+                return (
+                  <li className={styles.listItem} key={index}>
+                    <span className={styles.tokenIcon}>{asset.iconCircle}</span>
+                    <span>{asset.name}</span>
+                    <SvgImage className={classnames(
+                      styles.doneIcon,
+                      allowances[asset.symbol] ? styles.isAllowed : ''
+                    )} image={doneSvg}/>
+                  </li>
+                );
+              })
           }
-
-        </div>
+        </ul>
       </InstantFormWrapper>
 
     );
   }
 
-  private close = () => {
+  private onClose = () => {
     this.props.change({
       kind: InstantFormChangeKind.viewChange,
       view: ViewKind.account
