+++ src/utils/Variables.scss
@@ -14,7 +14,6 @@ $grey-darkest: #2F2F38; // components bg
 $grey-darken-darkest: #2B2B33; // panel border
 $very-dark-greyish-blue: #263238;
 $dark-greyish-blue: #555561;
-$yet-another-dark-grayish-blue: #6D6D74; //used in disabled done icon in token allowances view
 $darker-greyish-blue: #676773;
 $dark-gray-blue: #828287;
 $lighter-dark-gray-blue: #8D8D96;
