+++ src/utils/tooltip/Tooltip.scss
@@ -68,10 +68,4 @@
       fill: rgba(255, 255, 255, 0.9);
     }
   }
-
-  &.softCyan {
-    :global(.warning-icon) {
-      fill: $soft-cyan;
-    }
-  }
 }
