+++ src/utils/tooltip/Tooltip.scss.d.ts
@@ -1,4 +1,3 @@
 export const warningIcon: string;
 export const grey: string;
 export const white: string;
-export const softCyan: string;
